backend-fastapi/requirements.txt
fastapi
sqlmodel
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pymysql
requests
email-validator
cloudinary
python-multipart